cache: switch to wyhash for cache tables
This is dramatically faster on files with large drawings.

+++ libass/ass_cache.h
@@ -26,7 +26,7 @@
 #include "ass_bitmap.h"
 
 typedef struct cache Cache;
-typedef uint32_t ass_hashcode;
+typedef uint64_t ass_hashcode;
 
 // cache values
